Fix compilation warning in spdlog/fmt
fmtlib version used by spdlog doesn't like automatic enum conversions.
See:

https: //github.com/gabime/spdlog/issues/2560
https: //github.com/fmtlib/fmt/issues/3183
Reviewed-on: https://codereview.kdab.com/c/kdab/mecaps/+/134061

+++ src/network_access_manager/abstract_transfer_handle.cpp
@@ -93,7 +93,7 @@ void AbstractTransferHandle::transferDoneCallback(CURLcode result)
 	transferDoneCallbackImpl(result);
 
 	if (result != CURLcode::CURLE_OK) {
-		spdlog::error("curl transfer finished with code {} ({}) {}", result, curl_easy_strerror(result), error());
+		spdlog::error("curl transfer finished with code {} ({}) {}", static_cast<int>(result), curl_easy_strerror(result), error());
 	}
 
 	finished.emit((int)result);
